+++ packages/ecs-winston-format/examples/error.js
@@ -18,7 +18,7 @@
 'use strict'
 
 const winston = require('winston')
-const ecsFormat = require('../') // @elastic/ecs-winston-format
+const { ecsFormat } = require('../') // @elastic/ecs-winston-format
 
 const logger = winston.createLogger({
   level: 'info',